Move modules to single package mockServer
